No more failing to load wild pokemon
Before if you edited wild pokemon, you could no longer edit them if you exit and reloaded as they would all be blank. After rebuilding the code based off of LGPE editor it no longer fails to load! I also started to rework the gui to be similar to the LGPE editor, although that might be scrapped. Why are the methods between the two editors so different?