Fixed compatible TRs reading incorrectly. (#328)
* Fixed SWSH not gathering TR learnsets correctly

TR compatibility for each pokemon had been saving to the same list as TM compatibility, and for some reason it was not working as intended.
After finishing, and seeing that it built and ran, I tested the impacted features, and confirmed it is working.

* Remove Commented/Unused Section

PersonalDumperSWSH was previously commented, and is now entirely removed from PersonalDumper.cs

* Add case for TRs to SetIMovesInfo function

Changed IMovesInfo.cs
Added extra if for SWSH to account for TRs.

* Add case for SWSH to SetPersonalInfo function

new if statement in SetPersonalInfo from IPersonalInfo.cs that allows setting TRs.