Minor refactor
Rename all the interfaces to give them a more sensible name
Rename the Set<InterfaceName> extension methods to Import<InterfaceName>, since evenI got confused about it's purpose for a while. It's supposed to be used by the data importer to fill in missing properties from older generation data. This means it should actually not set every property that is part of the interface and 'set' is therefor a confusing name.
Update PokeDataUI to fix layout of new property names