Enhance PropertyGrid to handle IList<T> properties
PropertyGrid is an old control that doesn't acknowledge generic IList<T>, only IList (which this[index] returns object, not T)
so you have to give it handlers

FlatSharp can only generate IList<T>, IReadOnlyList<T>, (and some cases Memory/ReadOnlyMemory<T>, and IIndexedVector). So for this case, just implement IList<T> converter as it's the only one we really would like to use.